# backend/contracts/openapi.yaml
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewCorrection": {
        "additionalProperties": false,
        "properties": {
          "client_event_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Event Id",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "status": {
            "enum": [
              "known",
              "unknown",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "client_event_id",
          "expected_version"
        ],
        "title": "ReviewCorrection",
        "type": "object"
      },
      "ReviewCreate": {
        "additionalProperties": false,
        "properties": {
          "client_event_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Event Id",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "maximum": 86400000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "source": {
            "const": "quick_review",
            "default": "quick_review",
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "known",
              "unknown",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "word_id": {
            "exclusiveMinimum": 0.0,
            "title": "Word Id",
            "type": "integer"
          }
        },
        "required": [
          "word_id",
          "status",
          "client_event_id"
        ],
        "title": "ReviewCreate",
        "type": "object"
      },
...
      "TodayReviewCounts": {
        "additionalProperties": false,
        "properties": {
          "known": {
            "title": "Known",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "unknown": {
            "title": "Unknown",
            "type": "integer"
          }
        },
        "required": [
          "known",
          "unknown",
          "skipped",
          "total"
        ],
        "title": "TodayReviewCounts",
        "type": "object"
      },
      "TodayReviewData": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/TodayReviewCounts"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TodayReviewItem"
            },
            "title": "Items",
            "type": "array"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "date",
          "timezone",
          "counts",
          "items"
        ],
        "title": "TodayReviewData",
        "type": "object"
      },
      "TodayReviewItem": {
        "additionalProperties": false,
        "properties": {
          "cn_meaning": {
            "title": "Cn Meaning",
            "type": "string"
          },
          "en_word": {
            "title": "En Word",
            "type": "string"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonetic"
          },
          "review_id": {
            "title": "Review Id",
            "type": "integer"
          },
          "reviewed_at": {
            "title": "Reviewed At",
            "type": "string"
          },
          "round_id": {
            "title": "Round Id",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "integer"
          },
          "session_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Title"
          },
          "status": {
            "enum": [
              "known",
              "unknown",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          },
          "word_id": {
            "title": "Word Id",
            "type": "integer"
          }
        },
        "required": [
          "review_id",
          "round_id",
          "session_id",
          "session_title",
          "word_id",
          "en_word",
          "phonetic",
          "cn_meaning",
          "status",
          "reviewed_at"
        ],
        "title": "TodayReviewItem",
        "type": "object"
      },
      "TodayReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": "OK",
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/TodayReviewData"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "meta",
          "request_id"
        ],
        "title": "TodayReviewResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque",
        "description": "Managed high-entropy bearer token. Required scopes are listed in x-required-scopes because this is not an OAuth token endpoint.",
        "scheme": "bearer",
        "type": "http",
        "x-available-scopes": [
          "practice:generate",
          "practice:read",
          "practice:write",
          "reviews:read",
          "reviews:write",
          "words:export",
          "words:read",
          "words:write"
        ]
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "TrustedProxyUser": {
        "description": "Only accepted from TRUSTED_PROXY_CIDRS after the proxy strips client-supplied headers.",
        "in": "header",
        "name": "X-Forwarded-User",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-user word memory API. Browser requests use a trusted reverse-proxy identity; external Skills use scoped bearer tokens.",
    "title": "Word Memory Assistant API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "list_reviews_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "word_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Word Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "actor_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Type"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "round_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Round Id"
            }
          },
          {
            "in": "query",
            "name": "reviewed_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reviewed From"
            }
          },
          {
            "in": "query",
            "name": "reviewed_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reviewed To"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "reviewed_at_desc",
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "List Reviews",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:read"
        ]
      },
      "post": {
        "operationId": "create_review_api_v1_reviews_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Create Review",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:write",
          "words:read"
        ]
      }
    },
    "/api/v1/reviews/today": {
      "get": {
        "operationId": "today_online_reviews_api_v1_reviews_today_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Today Online Reviews",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:write",
          "practice:read"
        ]
      }
    },
    "/api/v1/reviews/{review_id}": {
      "patch": {
        "operationId": "patch_review_api_v1_reviews__review_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCorrection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Patch Review",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:write"
        ]
      }
    },
    "/api/v1/stats/summary": {
      "get": {
        "operationId": "stats_summary_api_v1_stats_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Stats Summary",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:read"
        ]
      }
    },
...
    "/api/v1/words/{word_id}/stats": {
      "get": {
        "operationId": "word_stats_api_v1_words__word_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "word_id",
            "required": true,
            "schema": {
              "title": "Word Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Word Stats",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "words:read",
          "reviews:read"
        ]
      }
    }
  }
}